requirements.txt
praw
pandas
transformers
torch
matplotlib
scikit-learn
nltk
opencv-python
python-dotenv
tqdm
tensorflow
